Create InAppMessages Namespace
Remove In-App Messages methods from OneSignalPlugin class, update method names, update iOS header and implementation files
